Separate the test and add new component object
