+++ e2e/Pages/DisasterTypeComponent.ts
@@ -32,19 +32,6 @@ class DisasterTypeComponent extends DashboardPage {
     await expect(this.droughtIcon).toBeVisible();
     await expect(this.topBar).toContainText(currentDateTime);
   }
-
-  async timlineElementsAreVisible() {
-    await this.page.waitForLoadState('domcontentloaded');
-    await this.page.waitForSelector('[data-testid="timeline-button"]');
-
-    const timelinePeriods = this.timeline;
-    const count = await timelinePeriods.count();
-
-    expect(count).toBeGreaterThan(0);
-    for (let i = 0; i < count; i++) {
-      await expect(timelinePeriods.nth(i)).toBeVisible();
-    }
-  }
 }
 
 export default DisasterTypeComponent;
