+++ e2e/tests/Timeline/TimelineComponentPresentNotrigger.spec.ts
@@ -1,6 +1,6 @@
 import { test } from '@playwright/test';
 import DashboardPage from 'Pages/DashboardPage';
-import DisasterTypeComponent from 'Pages/DisasterTypeComponent';
+import TimelineComponent from 'Pages/TimelineComponent';
 import UserStateComponent from 'Pages/UserStateComponent';
 import { qase } from 'playwright-qase-reporter';
 import { NoTriggerDataSet } from 'testData/testData.enum';
@@ -35,20 +35,16 @@ test.beforeEach(async ({ page }) => {
   );
 });
 
-test(
-  qase(4, 'All Disaster Type elements are present in no-trigger mode'),
-  async ({ page }) => {
-    const dashboard = new DashboardPage(page);
-    const userState = new UserStateComponent(page);
-    const disasterType = new DisasterTypeComponent(page);
+test(qase(15, 'Timeline present in no trigger mode'), async ({ page }) => {
+  const dashboard = new DashboardPage(page);
+  const userState = new UserStateComponent(page);
+  const timeline = new TimelineComponent(page);
 
-    // Navigate to disaster type the data was mocked for
-    await dashboard.navigateToFloodDisasterType();
-    // Assertions
-    await userState.headerComponentIsVisible({
-      countryName: NoTriggerDataSet.CountryName,
-    });
-    await disasterType.topBarComponentIsVisible();
-    await disasterType.allDisasterTypeElementsArePresent();
-  },
-);
+  // Navigate to disaster type the data was mocked for
+  await dashboard.navigateToFloodDisasterType();
+  // Assertions
+  await userState.headerComponentIsVisible({
+    countryName: NoTriggerDataSet.CountryName,
+  });
+  await timeline.timlineElementsAreVisible();
+});
